Move spec files into correct folder structure for unit tests
Missed that after rebasing
