+++ packages/web-app-files/tests/unit/components/Collaborators/Collaborator.spec.js
@@ -1,7 +1,7 @@
 import { createLocalVue, shallowMount } from '@vue/test-utils'
 import Vuex from 'vuex'
 import Collaborator from 'packages/web-app-files/src/components/Collaborators/Collaborator.vue'
-import stubs from '../../../../../tests/unit/stubs'
+import stubs from '../../../../../../tests/unit/stubs'
 import GetTextPlugin from 'vue-gettext'
 
 const localVue = createLocalVue()
